challenge/requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
numpy==1.26.3
outcome==1.3.0.post0
pandas==2.2.0
pyarrow==15.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wsproto==1.2.0
